remote_rom: change offset type to allow larger ROM content

+++ src/lib/remote_rom/backend/nic_ip/backend.cc
@@ -58,7 +58,7 @@ class Remote_rom::Packet_base : public Ethernet_frame, public Ipv4_packet
 		char         _module_name[MAX_NAME_LEN];   /* the ROM module name */
 		Type         _type;                        /* packet type */
 		uint32_t     _content_size;                /* ROM content size in bytes */
-		uint16_t     _offset;                      /* offset in bytes */
+		uint32_t     _offset;                      /* offset in bytes */
 		uint16_t     _payload_size;                /* payload size in bytes */
 
 		/*****************************************************
